+++ app/src/androidTest/java/com/mapbox/mapboxsdk/plugins/locationlayer/LocationLayerTest.java
@@ -6,8 +6,11 @@ import android.support.test.espresso.Espresso;
 import android.support.test.espresso.IdlingResourceTimeoutException;
 import android.support.test.espresso.UiController;
 import android.support.test.rule.ActivityTestRule;
+import android.support.test.rule.GrantPermissionRule;
 import android.support.test.runner.AndroidJUnit4;
 
+import com.mapbox.mapboxsdk.camera.CameraPosition;
+import com.mapbox.mapboxsdk.camera.CameraUpdateFactory;
 import com.mapbox.mapboxsdk.constants.Style;
 import com.mapbox.mapboxsdk.maps.MapboxMap;
 import com.mapbox.mapboxsdk.plugins.locationlayer.modes.RenderMode;
@@ -17,6 +20,7 @@ import com.mapbox.mapboxsdk.style.layers.SymbolLayer;
 import com.mapbox.mapboxsdk.utils.OnMapReadyIdlingResource;
 
 import org.junit.After;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -35,7 +39,6 @@ import static com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerConstants.
 import static com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerConstants.FOREGROUND_LAYER;
 import static com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerConstants.FOREGROUND_STALE_ICON;
 import static com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerConstants.LOCATION_SOURCE;
-import static com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerConstants.NAVIGATION_LAYER;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.not;
 import static org.junit.Assert.assertEquals;
@@ -43,11 +46,13 @@ import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
 @RunWith(AndroidJUnit4.class)
-@SuppressWarnings( {"MissingPermission"})
 public class LocationLayerTest {
 
   @Rule
   public ActivityTestRule<LocationLayerModesActivity> rule = new ActivityTestRule<>(LocationLayerModesActivity.class);
+
+  @Rule
+  public GrantPermissionRule permissionRule = GrantPermissionRule.grant(android.Manifest.permission.ACCESS_FINE_LOCATION);
 
   private OnMapReadyIdlingResource idlingResource;
   private LocationLayerPlugin locationLayerPlugin;
@@ -86,7 +91,7 @@ public class LocationLayerTest {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController, Context context) {
-        locationLayerPlugin.setLocationLayerMode(RenderMode.NORMAL);
+        locationLayerPlugin.setRenderMode(RenderMode.NORMAL);
         assertTrue(mapboxMap.getSource(LOCATION_SOURCE) != null);
       }
     });
@@ -98,7 +103,7 @@ public class LocationLayerTest {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController, Context context) {
-        locationLayerPlugin.setLocationLayerMode(RenderMode.NORMAL);
+        locationLayerPlugin.setRenderMode(RenderMode.NORMAL);
         assertTrue(mapboxMap.getLayer(ACCURACY_LAYER) != null);
         assertTrue(mapboxMap.getLayer(BACKGROUND_LAYER) != null);
         assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER) != null);
@@ -112,7 +117,7 @@ public class LocationLayerTest {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController, Context context) {
-        locationLayerPlugin.setLocationLayerMode(RenderMode.COMPASS);
+        locationLayerPlugin.setRenderMode(RenderMode.COMPASS);
         assertTrue(mapboxMap.getLayer(ACCURACY_LAYER) != null);
         assertTrue(mapboxMap.getLayer(BACKGROUND_LAYER) != null);
         assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER) != null);
@@ -127,8 +132,8 @@ public class LocationLayerTest {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController, Context context) {
-        locationLayerPlugin.setLocationLayerMode(RenderMode.COMPASS);
-        assertTrue(mapboxMap.getLayer(NAVIGATION_LAYER) != null);
+        locationLayerPlugin.setRenderMode(RenderMode.COMPASS);
+        assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER) != null);
       }
     });
   }
@@ -139,7 +144,7 @@ public class LocationLayerTest {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController, Context context) {
-        locationLayerPlugin.setLocationLayerMode(RenderMode.NORMAL);
+        locationLayerPlugin.setRenderMode(RenderMode.NORMAL);
         assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER) != null);
         locationLayerPlugin.setLocationLayerEnabled(false);
         assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER).getVisibility().getValue()
@@ -154,11 +159,11 @@ public class LocationLayerTest {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController, Context context) {
-        locationLayerPlugin.setLocationLayerMode(RenderMode.NORMAL);
+        locationLayerPlugin.setRenderMode(RenderMode.NORMAL);
         assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER) != null);
         mapboxMap.setStyleUrl(Style.SATELLITE);
         uiController.loopMainThreadForAtLeast(500);
-        assertEquals(locationLayerPlugin.getLocationLayerMode(), RenderMode.NORMAL);
+        assertEquals(locationLayerPlugin.getRenderMode(), RenderMode.NORMAL);
         assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER) != null);
         assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER).getVisibility().getValue()
           .equals(Property.VISIBLE));
@@ -176,11 +181,11 @@ public class LocationLayerTest {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController, Context context) {
-        locationLayerPlugin.setLocationLayerMode(RenderMode.NORMAL);
+        locationLayerPlugin.setRenderMode(RenderMode.NORMAL);
         SymbolLayer symbolLayer = mapboxMap.getLayerAs(FOREGROUND_LAYER);
         assert symbolLayer != null;
         assertThat(symbolLayer.getIconImage().getValue(), equalTo(FOREGROUND_ICON));
-        locationLayerPlugin.applyStyle(context, LocationLayerOptions.builder(context).staleStateTimeout(400).build());
+        locationLayerPlugin.applyStyle(LocationLayerOptions.builder(context).staleStateTimeout(400).build());
         locationLayerPlugin.forceLocationUpdate(location);
         uiController.loopMainThreadForAtLeast(500);
         assertThat(symbolLayer.getIconImage().getValue(), equalTo(FOREGROUND_STALE_ICON));
@@ -194,15 +199,28 @@ public class LocationLayerTest {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController, Context context) {
-        locationLayerPlugin.setLocationLayerMode(RenderMode.NORMAL);
-        locationLayerPlugin.applyStyle(context, LocationLayerOptions.builder(context).staleStateTimeout(100).build());
+        locationLayerPlugin.setRenderMode(RenderMode.NORMAL);
+        locationLayerPlugin.applyStyle(LocationLayerOptions.builder(context).staleStateTimeout(100).build());
         locationLayerPlugin.forceLocationUpdate(location);
         uiController.loopMainThreadForAtLeast(200);
         rule.getActivity().toggleStyle();
         SymbolLayer symbolLayer = mapboxMap.getLayerAs(FOREGROUND_LAYER);
         assert symbolLayer != null;
         assertThat(symbolLayer.getIconImage().getValue(), not(FOREGROUND_ICON));
-      }});
+      }
+    });
+  }
+
+  @Test
+  public void whenMapCameraInitializesTilted_iconsGetPlacedWithCorrectOffset() throws Exception {
+    executeLocationLayerTest((locationLayerPlugin, mapboxMap, uiController, context) -> {
+      mapboxMap.moveCamera(CameraUpdateFactory.newCameraPosition(new CameraPosition.Builder().tilt(60).build()));
+      locationLayerPlugin.setRenderMode(RenderMode.NORMAL);
+      locationLayerPlugin.forceLocationUpdate(location);
+      SymbolLayer layer = mapboxMap.getLayerAs(FOREGROUND_LAYER);
+      Float[] value = layer.getIconOffset().getValue();
+      Assert.assertEquals((-0.05 * 60), value[1], 0.1);
+    });
   }
 
   @After