+++ plugin-places/src/main/java/com/mapbox/mapboxsdk/plugins/places/autocomplete/data/dao/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Contains the Data Access Objects where we define the search history database interactions.
+ *
+ * @since 0.1.0
+ */
+package com.mapbox.mapboxsdk.plugins.places.autocomplete.data.dao;