+++ plugin-places/src/main/java/com/mapbox/mapboxsdk/plugins/places/autocomplete/ui/ResultClickCallback.java
@@ -1,0 +1,7 @@
+package com.mapbox.mapboxsdk.plugins.places.autocomplete.ui;
+
+import com.mapbox.api.geocoding.v5.models.CarmenFeature;
+
+interface ResultClickCallback {
+  void onClick(CarmenFeature carmenFeature);
+}