+++ plugin-places/src/main/java/com/mapbox/plugins/places/autocomplete/data/converter/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Contains type converters specific to Room to convert java objects into primitive types for easier
- * database storing.
- *
- * @since 0.1.0
- */
-package com.mapbox.plugins.places.autocomplete.data.converter;