+++ plugin-places/src/main/java/com/mapbox/plugins/places/autocomplete/model/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Contains interfaces which model a classes object methods.
- *
- * @since 0.1.0
- */
-package com.mapbox.plugins.places.autocomplete.model;