+++ plugin-places/src/main/java/com/mapbox/plugins/places/autocomplete/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Contains all the classes which can be useful between all other packages.
- *
- * @since 0.1.0
- */
-package com.mapbox.plugins.places.autocomplete;