+++ plugin-places/src/main/java/com/mapbox/plugins/places/autocomplete/ui/ResultClickCallback.java
@@ -1,7 +1,0 @@
-package com.mapbox.plugins.places.autocomplete.ui;
-
-import com.mapbox.api.geocoding.v5.models.CarmenFeature;
-
-interface ResultClickCallback {
-  void onClick(CarmenFeature carmenFeature);
-}