+++ plugin-places/CHANGELOG.md
@@ -2,8 +2,18 @@
 
 Mapbox welcomes participation and contributions from everyone.
 
-### 0.2.0 - TBD
-- Increased the search view button sizes to help capture when a user clicks the button.
+### 0.2.2 - March 7, 2018
+- Package namespacing is now corrected to match other plugins [#336](https://github.com/mapbox/mapbox-plugins-android/pull/336)
+- Back button in Place Picker activity now calls `finish()` [#337](https://github.com/mapbox/mapbox-plugins-android/pull/337)
+- Adds missing Javadoc information for `PlaceOptions` class [#339](https://github.com/mapbox/mapbox-plugins-android/pull/339)
+
+### 0.2.1 - February 27, 2018
+- Remove location layer dep [#319](https://github.com/mapbox/mapbox-plugins-android/pull/319)
+
+### 0.2.0 - February 26, 2018
+- Place Picker added [#262](https://github.com/mapbox/mapbox-plugins-android/pull/262)
+- Updated geocoding dependency [#261](https://github.com/mapbox/mapbox-plugins-android/pull/261)
+- Increased the search view button sizes to help capture when a user clicks the button. [#240](https://github.com/mapbox/mapbox-plugins-android/pull/240)
 
 ### 0.1.0 - December 22, 2017
 - Initial release as a standalone package.